+++ test/test.tsx
@@ -107,20 +107,19 @@ import {
   WizardBody as WizardBodyNext,
   WizardFooter,
 } from "@patternfly/react-core/next";
-
-import { Td } from "@patternfly/react-table";
-
-const tdSelectTypeObj = { disable: true };
 import {
   Table,
   TableBody,
   TableHeader,
   TableProps,
+  Td,
+  Tr,
 } from "@patternfly/react-table";
 import { TableComposable } from "@patternfly/react-table";
 import { TableComposable as PFTable } from "@patternfly/react-table";
 
 
+const tdSelectTypeObj = { disable: true };
 //following type of import was causing errors for rules that checked specifiers before import package
 import foo from "Bar";
 
@@ -262,6 +261,7 @@ const alertVariantOption = AlertVariant.default;
   <Spinner isSVG />
   <Switch onChange={(foo) => handler(foo)} />
   <Table />
+  <Table rows={[]} />
   <TableBody />
   <TableComposable />
   <TableHeader />
@@ -270,6 +270,7 @@ const alertVariantOption = AlertVariant.default;
   <Toggle isPrimary onToggle={} />
   <ToggleGroupItem onChange={(foo, event) => handler(foo, event)} />
   <Tooltip reference />
+  <Tr isHoverable />
   <TreeView hasCheck />
   <Wizard />
   <Wizard mainAriaLabel />