+++ packages/eslint-plugin-pf-codemods/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.69.2](https://github.com/patternfly/pf-codemods/compare/@patternfly/eslint-plugin-pf-codemods@1.69.1...@patternfly/eslint-plugin-pf-codemods@1.69.2) (2023-05-16)
+
+
+### Bug Fixes
+
+* **Table:** updated isHoverable to isClickable ([#470](https://github.com/patternfly/pf-codemods/issues/470)) ([76a1c79](https://github.com/patternfly/pf-codemods/commit/76a1c79e62d1403a883354bb9c77bf64c65e0275))
+
+
+
+
+
+## [1.69.1](https://github.com/patternfly/pf-codemods/compare/@patternfly/eslint-plugin-pf-codemods@1.69.0...@patternfly/eslint-plugin-pf-codemods@1.69.1) (2023-05-15)
+
+**Note:** Version bump only for package @patternfly/eslint-plugin-pf-codemods
+
+
+
+
+
 # [1.69.0](https://github.com/patternfly/pf-codemods/compare/@patternfly/eslint-plugin-pf-codemods@1.68.0...@patternfly/eslint-plugin-pf-codemods@1.69.0) (2023-05-15)